V0.9 Add casts/etc. in vvp directory to remove Cygwin compile warnings.

+++ vvp/vpi_callback.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2001-2008 Stephen Williams (steve@icarus.com)
+ * Copyright (c) 2001-2009 Stephen Williams (steve@icarus.com)
  *
  *    This source code is free software; you can redistribute it
  *    and/or modify it in source code form under the terms of the GNU
@@ -259,7 +259,8 @@ static struct __vpiCallback* make_sync(p_cb_data data, bool readonly_flag)
 	      }
 
 	  default:
-	    fprintf(stderr, "Unsupported time type %d.\n", obj->cb_time.type);
+	    fprintf(stderr, "Unsupported time type %d.\n",
+	            (int)obj->cb_time.type);
 	    assert(0);
 	    break;
       }
@@ -288,7 +289,8 @@ static struct __vpiCallback* make_afterdelay(p_cb_data data, bool simtime_flag)
 	    break;
 
 	  default:
-	    fprintf(stderr, "Unsupported time type %d.\n", obj->cb_time.type);
+	    fprintf(stderr, "Unsupported time type %d.\n",
+	            (int)obj->cb_time.type);
 	    assert(0);
 	    tv = 0;
 	    break;
@@ -465,7 +467,7 @@ vpiHandle vpi_register_cb(p_cb_data data)
 	  default:
 	    fprintf(stderr, "vpi error: vpi_register_cb invalid or "
 		    "unsupported callback reason: %d\n",
-		    data->reason);
+		    (int)data->reason);
 	    break;
       }
 
@@ -510,7 +512,7 @@ void callback_execute(struct __vpiCallback*cur)
 	    break;
 	  default:
 	    fprintf(stderr, "Unsupported time format %d.\n",
-	            cur->cb_data.time->type);
+	            (int)cur->cb_data.time->type);
 	    assert(0);
 	    break;
       }
@@ -641,7 +643,7 @@ void vvp_fun_signal4::get_value(struct t_vpi_value*vp)
 	  default:
 	    fprintf(stderr, "vpi_callback: value "
 		    "format %d not supported (fun_signal)\n",
-		    vp->format);
+		    (int)vp->format);
       }
 }
 
@@ -656,7 +658,7 @@ void vvp_fun_signal8::get_value(struct t_vpi_value*vp)
 	  default:
 	    fprintf(stderr, "vpi_callback: value "
 		    "format %d not supported (fun_signal8)\n",
-		    vp->format);
+		    (int)vp->format);
       }
 }
 
@@ -716,6 +718,6 @@ void vvp_fun_signal_real::get_value(struct t_vpi_value*vp)
 	  default:
 	    fprintf(stderr, "vpi_callback: value "
 		    "format %d not supported (fun_signal_real)\n",
-		    vp->format);
+		    (int)vp->format);
       }
 }
